Exam Quixo/requirements.txt
absl-py==1.1.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
coloredlogs==15.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.4
libclang==14.0.1
Markdown==3.3.7
numpy
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame
pyparsing==3.0.9
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
termcolor==1.1.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
pillow